requirements.txt
TA-Lib
futu-api
click
